+++ core-api/src/modules/reports/reports.module.ts
@@ -6,14 +6,15 @@ import { ReportsService } from './reports.service';
 import { VulnerabilityReportService } from './services/vulnerability-report.service';
 import { SummaryReportService } from './services/summary-report.service';
 import { Vulnerability } from '@/modules/vulnerabilities/entities/vulnerability.entity';
-import { Asset } from '@/modules/assets/entities/assets.entity';
 import { Target } from '@/modules/targets/entities/target.entity';
 import { AssetService } from '@/modules/assets/entities/asset-services.entity';
 import { AssetTag } from '@/modules/assets/entities/asset-tags.entity';
+import { StatisticModule } from '@/modules/statistic/statistic.module';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Report, Vulnerability, Asset, Target, AssetService, AssetTag]),
+    TypeOrmModule.forFeature([Report, Vulnerability, Target, AssetService, AssetTag]),
+    StatisticModule,
   ],
   controllers: [ReportsController],
   providers: [ReportsService, VulnerabilityReportService, SummaryReportService],
