+++ core-api/src/modules/reports/reports.service.ts
@@ -248,6 +248,16 @@ export class ReportsService {
         return arr.join(separator);
       });
 
+      Handlebars.registerHelper('formatDate', (dateStr: string | Date | null) => {
+        if (!dateStr) return '-';
+        const date = new Date(dateStr);
+        if (isNaN(date.getTime())) return '-';
+        const day = String(date.getDate()).padStart(2, '0');
+        const month = String(date.getMonth() + 1).padStart(2, '0');
+        const year = date.getFullYear();
+        return `${day}/${month}/${year}`;
+      });
+
       this.handlebarsTemplate = Handlebars.compile<ReportData>(templateContent);
     } catch {
       // ignore
