+++ core-api/src/modules/reports/services/vulnerability-report.service.ts
@@ -2,9 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Vulnerability } from '@/modules/vulnerabilities/entities/vulnerability.entity';
-import { Asset } from '@/modules/assets/entities/assets.entity';
-import { Target } from '@/modules/targets/entities/target.entity';
 import { Severity, VulnerabilityAnalyzeStatus } from '@/common/enums/enum';
+import { StatisticService } from '@/modules/statistic/statistic.service';
 import type { VulnerabilityReportData } from '../types/vulnerability-report-data.type';
 
 @Injectable()
@@ -12,10 +11,7 @@ export class VulnerabilityReportService {
   constructor(
     @InjectRepository(Vulnerability)
     private readonly vulnRepo: Repository<Vulnerability>,
-    @InjectRepository(Asset)
-    private readonly assetRepo: Repository<Asset>,
-    @InjectRepository(Target)
-    private readonly targetRepo: Repository<Target>,
+    private readonly statisticService: StatisticService,
   ) {}
 
   async getVulnerabilityReportData(
@@ -41,15 +37,15 @@ export class VulnerabilityReportService {
       analyzeStats,
       targetStats,
     ] = await Promise.all([
-      this.getSeverityCounts(workspaceId, options),
+      this.statisticService.getSeverityCounts(workspaceId, options),
       this.getRecentVulnerabilities(workspaceId, options),
-      this.getTopAssetsWithVulnerabilities(workspaceId, options),
-      this.getVulnerabilityTrends(workspaceId, options),
+      this.statisticService.getTopAssetsWithVulnerabilities(workspaceId, options),
+      this.statisticService.getVulnerabilityTrends(workspaceId, options),
       this.getCriticalVulnerabilities(workspaceId, options),
       this.getAllVulnerabilities(workspaceId, options),
       this.getDismissedStats(workspaceId, options),
       this.getAnalyzeStats(workspaceId, options),
-      this.getTargetVulnerabilityStats(workspaceId, options),
+      this.statisticService.getTargetVulnerabilityStats(workspaceId, options),
     ]);
 
     const totalVulns = Object.values(severityCounts).reduce(
@@ -131,52 +127,6 @@ export class VulnerabilityReportService {
     };
   }
 
-  private async getSeverityCounts(
-    workspaceId: string,
-    options?: { startDate?: Date; endDate?: Date; targetIds?: string[] },
-  ): Promise<Record<Severity, number>> {
-    const qb = this.vulnRepo
-      .createQueryBuilder('v')
-      .leftJoin('v.asset', 'asset')
-      .leftJoin('asset.target', 'target')
-      .leftJoin('target.workspaceTargets', 'wt')
-      .leftJoin('wt.workspace', 'workspace')
-      .select('v.severity', 'severity')
-      .addSelect('COUNT(*)', 'count')
-      .where('workspace.id = :workspaceId', { workspaceId })
-      .andWhere('v.isArchived = :isArchived', { isArchived: false });
-
-    if (options?.startDate) {
-      qb.andWhere('v.createdAt >= :startDate', { startDate: options.startDate });
-    }
-    if (options?.endDate) {
-      qb.andWhere('v.createdAt <= :endDate', { endDate: options.endDate });
-    }
-    if (options?.targetIds?.length) {
-      qb.andWhere('target.id IN (:...targetIds)', {
-        targetIds: options.targetIds,
-      });
-    }
-
-    const results = await qb
-      .groupBy('v.severity')
-      .getRawMany<{ severity: Severity; count: string }>();
-
-    const counts = {
-      [Severity.CRITICAL]: 0,
-      [Severity.HIGH]: 0,
-      [Severity.MEDIUM]: 0,
-      [Severity.LOW]: 0,
-      [Severity.INFO]: 0,
-    };
-
-    for (const result of results) {
-      counts[result.severity] = parseInt(result.count, 10);
-    }
-
-    return counts;
-  }
-
   private async getRecentVulnerabilities(
     workspaceId: string,
     options?: { startDate?: Date; endDate?: Date; targetIds?: string[] },
@@ -237,166 +187,6 @@ export class VulnerabilityReportService {
       firstDetectedDate: v.firstDetectedDate?.toISOString() || null,
       lastSeenDate: v.lastSeenDate?.toISOString() || null,
     }));
-  }
-
-  private async getTopAssetsWithVulnerabilities(
-    workspaceId: string,
-    options?: { startDate?: Date; endDate?: Date; targetIds?: string[] },
-    limit = 10,
-  ) {
-    const qb = this.vulnRepo
-      .createQueryBuilder('v')
-      .leftJoin('v.asset', 'asset')
-      .leftJoin('asset.target', 'target')
-      .leftJoin('target.workspaceTargets', 'wt')
-      .leftJoin('wt.workspace', 'workspace')
-      .select('asset.value', 'assetValue')
-      .addSelect('COUNT(*)', 'totalCount')
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'critical' THEN 1 ELSE 0 END)`,
-        'critical',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'high' THEN 1 ELSE 0 END)`,
-        'high',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'medium' THEN 1 ELSE 0 END)`,
-        'medium',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'low' THEN 1 ELSE 0 END)`,
-        'low',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'info' THEN 1 ELSE 0 END)`,
-        'info',
-      )
-      .where('workspace.id = :workspaceId', { workspaceId })
-      .andWhere('v.isArchived = :isArchived', { isArchived: false })
-      .groupBy('asset.value')
-      .orderBy('"totalCount"', 'DESC')
-      .limit(limit);
-
-    if (options?.startDate) {
-      qb.andWhere('v.createdAt >= :startDate', { startDate: options.startDate });
-    }
-    if (options?.endDate) {
-      qb.andWhere('v.createdAt <= :endDate', { endDate: options.endDate });
-    }
-    if (options?.targetIds?.length) {
-      qb.andWhere('target.id IN (:...targetIds)', {
-        targetIds: options.targetIds,
-      });
-    }
-
-    const results = await qb.getRawMany<{
-      assetValue: string;
-      totalCount: string;
-      critical: string;
-      high: string;
-      medium: string;
-      low: string;
-      info: string;
-    }>();
-
-    return results.map((r) => ({
-      asset: r.assetValue,
-      total: parseInt(r.totalCount, 10),
-      critical: parseInt(r.critical, 10),
-      high: parseInt(r.high, 10),
-      medium: parseInt(r.medium, 10),
-      low: parseInt(r.low, 10),
-      info: parseInt(r.info, 10),
-    }));
-  }
-
-  private async getVulnerabilityTrends(
-    workspaceId: string,
-    options?: { startDate?: Date; endDate?: Date; targetIds?: string[] },
-  ) {
-    const now = new Date();
-    const last7Days = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-    const last30Days = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-
-    const [last7DaysData, last30DaysData] = await Promise.all([
-      this.getDailyVulnerabilityCounts(workspaceId, last7Days, now, options),
-      this.getDailyVulnerabilityCounts(workspaceId, last30Days, now, options),
-    ]);
-
-    const avgPerWeek =
-      last7DaysData.length > 0
-        ? last7DaysData.reduce((sum, d) => sum + d.count, 0) / 7
-        : 0;
-
-    // Calculate trend
-    const firstHalf = last30DaysData.slice(0, 15);
-    const secondHalf = last30DaysData.slice(15);
-    const firstAvg =
-      firstHalf.length > 0
-        ? firstHalf.reduce((sum, d) => sum + d.count, 0) / firstHalf.length
-        : 0;
-    const secondAvg =
-      secondHalf.length > 0
-        ? secondHalf.reduce((sum, d) => sum + d.count, 0) / secondHalf.length
-        : 0;
-
-    let trend: 'increasing' | 'decreasing' | 'stable' = 'stable';
-    if (secondAvg > firstAvg * 1.2) trend = 'increasing';
-    else if (secondAvg < firstAvg * 0.8) trend = 'decreasing';
-
-    return {
-      last7Days: last7DaysData.map((d) => d.count),
-      last30Days: last30DaysData.map((d) => d.count),
-      avgPerWeek: Math.round(avgPerWeek * 10) / 10,
-      trend,
-    };
-  }
-
-  private async getDailyVulnerabilityCounts(
-    workspaceId: string,
-    startDate: Date,
-    endDate: Date,
-    options?: { targetIds?: string[] },
-  ) {
-    const qb = this.vulnRepo
-      .createQueryBuilder('v')
-      .leftJoin('v.asset', 'asset')
-      .leftJoin('asset.target', 'target')
-      .leftJoin('target.workspaceTargets', 'wt')
-      .leftJoin('wt.workspace', 'workspace')
-      .select("TO_CHAR(v.createdAt, 'YYYY-MM-DD')", 'date')
-      .addSelect('COUNT(*)', 'count')
-      .where('workspace.id = :workspaceId', { workspaceId })
-      .andWhere('v.createdAt >= :startDate', { startDate })
-      .andWhere('v.createdAt <= :endDate', { endDate })
-      .andWhere('v.isArchived = :isArchived', { isArchived: false })
-      .groupBy("TO_CHAR(v.createdAt, 'YYYY-MM-DD')")
-      .orderBy('date', 'ASC');
-
-    if (options?.targetIds?.length) {
-      qb.andWhere('target.id IN (:...targetIds)', {
-        targetIds: options.targetIds,
-      });
-    }
-
-    const results = await qb.getRawMany<{ date: string; count: string }>();
-
-    // Fill in missing days with 0 counts
-    const dateMap = new Map(results.map((r) => [r.date, parseInt(r.count, 10)]));
-    const dates: { date: string; count: number }[] = [];
-
-    const current = new Date(startDate);
-    while (current <= endDate) {
-      const dateStr = current.toISOString().split('T')[0];
-      dates.push({
-        date: dateStr,
-        count: dateMap.get(dateStr) || 0,
-      });
-      current.setDate(current.getDate() + 1);
-    }
-
-    return dates;
   }
 
   private async getCriticalVulnerabilities(
@@ -615,80 +405,6 @@ export class VulnerabilityReportService {
     }
 
     return stats;
-  }
-
-  private async getTargetVulnerabilityStats(
-    workspaceId: string,
-    options?: { startDate?: Date; endDate?: Date; targetIds?: string[] },
-  ) {
-    const qb = this.targetRepo
-      .createQueryBuilder('t')
-      .leftJoin('t.workspaceTargets', 'wt')
-      .leftJoin('wt.workspace', 'workspace')
-      .leftJoin('t.assets', 'asset')
-      .leftJoin('asset.vulnerabilities', 'v')
-      .select('t.value', 'targetValue')
-      .addSelect('t.type', 'targetType')
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'critical' THEN 1 ELSE 0 END)`,
-        'critical',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'high' THEN 1 ELSE 0 END)`,
-        'high',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'medium' THEN 1 ELSE 0 END)`,
-        'medium',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'low' THEN 1 ELSE 0 END)`,
-        'low',
-      )
-      .addSelect(
-        `SUM(CASE WHEN v.severity = 'info' THEN 1 ELSE 0 END)`,
-        'info',
-      )
-      .addSelect('COUNT(v.id)', 'total')
-      .where('workspace.id = :workspaceId', { workspaceId })
-      .andWhere('v.isArchived = :isArchived', { isArchived: false })
-      .groupBy('t.id')
-      .addGroupBy('t.value')
-      .addGroupBy('t.type')
-      .having('COUNT(v.id) > 0')
-      .orderBy('total', 'DESC');
-
-    if (options?.startDate) {
-      qb.andWhere('v.createdAt >= :startDate', { startDate: options.startDate });
-    }
-    if (options?.endDate) {
-      qb.andWhere('v.createdAt <= :endDate', { endDate: options.endDate });
-    }
-    if (options?.targetIds?.length) {
-      qb.andWhere('t.id IN (:...targetIds)', { targetIds: options.targetIds });
-    }
-
-    const results = await qb.getRawMany<{
-      targetValue: string;
-      targetType: string;
-      critical: string;
-      high: string;
-      medium: string;
-      low: string;
-      info: string;
-      total: string;
-    }>();
-
-    return results.map((r) => ({
-      target: r.targetValue,
-      type: r.targetType,
-      critical: parseInt(r.critical, 10),
-      high: parseInt(r.high, 10),
-      medium: parseInt(r.medium, 10),
-      low: parseInt(r.low, 10),
-      info: parseInt(r.info, 10),
-      total: parseInt(r.total, 10),
-    }));
   }
 
   private async getAverageCvssScore(
